➗ Separación de todos los modelos de la clase Modelos.cs por separado para mejor estructura del código.
